Update ICS calendar sending flow:
* Move all related operations (helpers and thunks) to 'store/actions/icsOperations' (from both the 'AddToCalendar' component and 'bookingOperations')
* Replace internal (Material UI implemented) modal with 'useModal' flow and ActionDialog (prompting email input)
