+++ packages/client/src/features/modal/components/SendICSDialog/index.ts
@@ -1,0 +1,3 @@
+import SendICSDialog from "./SendICSDialog";
+
+export default SendICSDialog;
